+++ felicity/Backend/app/index.js
@@ -12,14 +12,9 @@ app.use(express.json());
 app.use(require("./patientLogin/router"));
 app.use(require("./posts/router"));
 app.use(require("./schedules/router"));
-// 1
-const plogin = require("./patientLogin/login_model")
-const dlogin = require("./doctorLogin/login_model");
-const jwt = require("jsonwebtoken");
-require('dotenv').config({path: './app/.env'})
-// const cookieParser = require('cookie-parser')
-// app.use(cookieParser())
-//
+
+const login = require("./login.js")
+
 app.get("/audio", (req, res) => {
     console.log(req.body);
 });
@@ -29,75 +24,22 @@ const server = app.listen(config.express.port, () => {
     console.log(`Server running on Port ${config.express.port}`);
 });
 
-// 2
-// require('crypto').randomBytes(64).toString('hex') --> 난수생성
-Object.prototype.getByIndex = function(index) {
-    return this[Object.keys(this)[index]];
-  };
-
-function authenticateToken(req, res, next) {
-    const authHeader = req.headers['authorization']
-    const token = authHeader && authHeader.split(' ')[1]
-    if (token == null) return console.log(err) //res.response(403)
-
-    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, payload) => {
-        if (err) return console.log(err) //res.response(403)
-        req.payload = payload
-        next()
-    })
-}
-// Athenticating using cookie
-// function authenticateTokenCookie(req, res, next){
-//     const accessToken = req.cookies.jwt
-//     if(!accessToken){
-//         return res.response(403)
-//     }
-//     try{
-//         const payload = jwt.verify(accessToken, process.env.ACCESS_TOKEN_SECRET)
-//         next()
-//     }
-//     catch(e){
-//         return res.response(403)
-//     }
-// }
-
-function generateAccessToken(payload){
-    return jwt.sign(payload, process.env.ACCESS_TOKEN_SECRET, {algorithm: "HS256", expiresIn: '10m'})
-}
-
+//login using jwt
+//doctor login
 app.post('/dlogin', (req, res) => {
     const email = req.body.email
     const password = req.body.password
-    dlogin.doctorLogin([email, password], (err, result) => {
-        if (err) console.log(err);
-        // res.json(result);
-        const doctorID = result[0].getByIndex(0)
-        console.log(doctorID)
-        const payload = {email:email}
-        const accessToken = generateAccessToken(payload)
-        res.json({accessToken:accessToken, doctorID:doctorID})
-        //res.cookie("jwt", accessToken)
-        //res.send()
-    });
+    const result = login.docLogin([email, password])
+    res.json({accessToken:result[1], doctorID:result[0]})
 })
 
+//patient login
 app.post('/plogin', (req, res) => {
     const email = req.body.email
     const password = req.body.password
-    plogin.patientLogin([email, password], (err, result) =>{
-        if (err) console.log(err);
-        // res.json(result);
-        const patientID = result[0].getByIndex(0)
-        console.log(patientID)
-        const payload = {email:eamil}
-        const accessToken = generateAccessToken(payload)
-        res.json({accessToken:accessToken, doctorID:doctorID})
-        //res.cooke("jwt", accessToken)
-        //res.send()
-    })
-    
+    const result = login.paLogin([email, password])
+    res.json({accessToken:result[1], doctorID:result[0]})
 })
-//
 
 
 const io = socket(server, {
